Refactor and restructure the code.

+++ include/utils.h
@@ -41,7 +41,7 @@ public:
     }
 
     void pushFreeChunk(T elementForPush) {
-        for (int i = 0; i < size_ ; i++) {
+        for (size_t i = 0; i < size_ ; i++) {
             if (container_[i] == elementForPush) {
                 free_chunks_[i] = true;
                 count_free_chunks_++;
